observaciones
observaciones personas

+++ src/main/java/com/app/ista/repository/ObservacionesPersonaRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.mongodb.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ObservacionesPersonaRepository extends MongoRepository<ObservacionesPersona, Long>{
+public interface ObservacionesPersonaRepository extends MongoRepository<ObservacionesPersona, Integer>{
     @Query(value = "{},{idObservacionesPersona:1}", sort = "{_id: -1}")
     List<ObservacionesPersona> id();
     List<ObservacionesPersona> findByCedulaPersona(String cedulaPersona);
